Refactor: extract runtime logic from codegen into library functions
Move boot protocol logic out of codegen quote!{} blocks into testable,
debuggable library functions in fstart-capabilities.  Codegen now generates
wiring calls instead of inlining entire boot sequences.

Library extractions:
- fit::load_fit_components() — FIT runtime parsing, config resolution,
  kernel/ramdisk copy to load addresses (replaces ~100 lines of generated code)
- scan_anchor_in_media() — memory-mapped FFS anchor scanning
- read_anchor_at_offset() — block device FFS anchor read + magic verify

Module split of capabilities.rs (2163 lines) into:
- capabilities/mod.rs (838) — simple capabilities + shared helpers
- capabilities/payload.rs (573) — PayloadLoad (Linux, UEFI, FIT)
- capabilities/acpi.rs (259) — AcpiPrepare + platform ACPI
- capabilities/smbios.rs (183) — SmBiosPrepare
- capabilities/stage_load.rs (257) — LoadNextStage + anchor scan

Also replace unstable SyncUnsafeCell with a stable #[repr(transparent)]
wrapper struct, removing the #![feature(sync_unsafe_cell)] dependency.
